+++ docs/pages/changelog.rst
@@ -11,6 +11,7 @@ Added
 
 - The tree plugin can now filter content based on a search field.
 - In debugMode, items from the tree can be added to the Console.
+- Added action to liveDisplays to toggle between automatically or manually scale the x axis.
 
 Fixed
 ~~~~~